refactor: move addMissingLayouts service
